refactor: 스크롤버튼 컴포넌트 분리, 스크롤 버튼 기능 수정
