+++ src/app/thems.scss
@@ -6,6 +6,12 @@ $bg-color-3: #da3017;
 $bg-color-4: #eee;
 $bg-color-5: #3c3c3c;
 
+$hover-color-3: #ce3017;
+
+html {
+  --color-main-bg-hover: #{$hover-color-3};
+}
+
 html[data-theme='light'] {
   --color-text: #{$text-color-1};
   --color-background: #{$bg-color-1};
@@ -19,3 +25,4 @@ html[data-theme='dark'] {
   --color-main-bg: #{$bg-color-2};
   --color-videoInfo-bg: #{$bg-color-5};
 }
+
