+++ src/components/ScrollBtn.module.scss
@@ -1,0 +1,23 @@
+.scrollBtn {
+  width: 3.75rem;
+  height: 3.75rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: fixed;
+  bottom: 3rem;
+  right: 3rem;
+  border-radius: 50%;
+  background-color: var(--color-main-bg);
+  transition: .3s ease;
+
+  &:hover {
+    background-color: var(--color-main-bg-hover);
+  }
+
+  .arrowUpIcon, .arrowDownIcon {
+    font-size: 1.5rem;
+    color: #fff;
+  }
+
+}
